+++ src/interfaces/actions/Action.sol
@@ -3,5 +3,5 @@
 pragma solidity ^0.8.0;
 
 interface IAction {
-    function execute(bytes calldata) external returns (uint256);
+    function execute(bytes calldata) external returns (uint);
 }
